fixed prev task v2 & created mock file
